requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
sqlalchemy
langchain
langchain-community
langchain-openai
chromadb
httpx
sentence-transformers
pymupdf
beautifulsoup4
pydantic-settings
openpyxl
langchain-ollama
langchain-chroma
langchain-huggingface
beautifulsoup4
unstructured

